requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yfinance==0.2.28
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
